messaging: Move SnackBarService to new MessagingModule
